advancedb_project/backend/requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
bcrypt==3.2.0
python-dotenv==0.19.0
python-jose==3.3.0
gunicorn==20.1.0
supabase==1.0.3
supabase-py==0.0.2